+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [Unreleased]
+#### Changed
+- Now including offer fees to account's balance on invest form
+
+#### Removed
+- Fixed fee from invest form
+
 ## [1.7.0-rc.0] - 2019-05-21
 #### Added
 - New sale whitelist support: